RIA-7174 Refactor package location
